+++ packages/mobile/src/page/Article/Detail/style.module.scss
